llm_assignment/requirements.txt
flask
python-dotenv
watchdog
ctransformers
flask-cors
